requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
aiofiles>=23.2.1
google-genai>=1.0.0
python-dotenv>=1.0.0
pypdf>=4.2.0
pdfplumber>=0.11.4
python-docx>=1.1.0
pydantic>=2.7.0
slowapi>=0.1.9
numpy>=1.26.0
httpx>=0.27.0
